# docs/api/openapi.json
{
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "OrganizeActivityOut": {
        "description": "Recent churn in a folder (ADR-021 Phase 3 watch trigger) — a hint to re-organise.\n\nRead-only summary off the incremental change feed (``change_log``): how many files were\ncreated/modified/deleted under ``path`` within the window. ``suggests_reorganise`` is True when\nnew or changed files appeared — the UI surfaces it as a \"re-organise?\" nudge; nothing is ever\nauto-applied (a human still approves every plan).",
        "properties": {
          "capped": {
            "title": "Capped",
            "type": "boolean"
          },
          "created": {
            "title": "Created",
            "type": "integer"
          },
          "deleted": {
            "title": "Deleted",
            "type": "integer"
          },
          "modified": {
            "title": "Modified",
            "type": "integer"
          },
          "path": {
            "title": "Path",
            "type": "string"
          },
          "since_hours": {
            "title": "Since Hours",
            "type": "integer"
          },
          "suggests_reorganise": {
            "title": "Suggests Reorganise",
            "type": "boolean"
          },
          "volume_id": {
            "title": "Volume Id",
            "type": "integer"
          }
        },
        "required": [
          "volume_id",
          "path",
          "since_hours",
          "created",
          "modified",
          "deleted",
          "capped",
          "suggests_reorganise"
        ],
        "title": "OrganizeActivityOut",
        "type": "object"
      },
      "OrganizeItemOut": {
        "description": "One file's proposed disposition after server-side validation/clamping.\n\n``status`` is ``move`` (relocate/rename to ``proposed_relpath``), ``keep`` (already placed\nwell), or ``rejected`` (the model's target escaped the root, was a bad name, or collided —\nnever moved). ``proposed_relpath`` is relative to the requested folder.",
        "properties": {
          "current_name": {
            "title": "Current Name",
            "type": "string"
          },
          "current_path": {
            "title": "Current Path",
            "type": "string"
          },
          "entry_id": {
            "title": "Entry Id",
            "type": "integer"
          },
          "proposed_name": {
            "title": "Proposed Name",
            "type": "string"
          },
          "proposed_relpath": {
            "title": "Proposed Relpath",
            "type": "string"
          },
          "reason": {
            "title": "Reason",
            "type": "string"
          },
          "status": {
            "title": "Status",
            "type": "string"
          }
        },
        "required": [
          "entry_id",
          "current_path",
          "current_name",
          "proposed_relpath",
          "proposed_name",
          "reason",
          "status"
        ],
        "title": "OrganizeItemOut",
        "type": "object"
      },
      "OrganizeMoveIn": {
        "description": "One operator-approved relocation: a catalogue entry + its in-root target (re-clamped).",
        "properties": {
          "dest_rel": {
            "maxLength": 4096,
            "minLength": 1,
            "title": "Dest Rel",
            "type": "string"
          },
          "entry_id": {
            "minimum": 1.0,
            "title": "Entry Id",
            "type": "integer"
          }
        },
        "required": [
          "entry_id",
          "dest_rel"
        ],
        "title": "OrganizeMoveIn",
        "type": "object"
      },
      "OrganizePlanItemOut": {
        "properties": {
          "dest_rel": {
            "title": "Dest Rel",
            "type": "string"
          },
          "entry_id": {
            "title": "Entry Id",
            "type": "integer"
          },
          "path": {
            "title": "Path",
            "type": "string"
          }
        },
        "required": [
          "entry_id",
          "path",
          "dest_rel"
        ],
        "title": "OrganizePlanItemOut",
        "type": "object"
      },
      "OrganizePlanOut": {
        "description": "A built, persisted MOVE plan — feed ``plan_id`` to the remediation dry-run/execute spine.",
        "properties": {
          "blast_count": {
            "title": "Blast Count",
            "type": "integer"
          },
          "host_id": {
            "title": "Host Id",
            "type": "string"
          },
          "items": {
            "items": {
              "$ref": "#/components/schemas/OrganizePlanItemOut"
            },
            "title": "Items",
            "type": "array"
          },
          "move_root": {
            "title": "Move Root",
            "type": "string"
          },
          "plan_id": {
            "title": "Plan Id",
            "type": "string"
          },
          "reclaimable_bytes": {
            "title": "Reclaimable Bytes",
            "type": "integer"
          },
          "status": {
            "title": "Status",
            "type": "string"
          }
        },
        "required": [
          "plan_id",
          "move_root",
          "host_id",
          "blast_count",
          "reclaimable_bytes",
          "status",
          "items"
        ],
        "title": "OrganizePlanOut",
        "type": "object"
      },
      "OrganizePlanRequest": {
        "description": "Build a reversible MOVE plan from a reviewed proposal's approved subset (ADR-023).\n\nThe client sends only ``entry_id`` + the approved ``dest_rel`` per move; the server pulls the\ndrift anchor (inode/size/hash) and source path from the catalogue and re-clamps every target to\n``path`` (AR-0012 — the operator chose the destinations, the server owns the rest).",
        "properties": {
          "idempotency_key": {
            "anyOf": [
              {
                "maxLength": 128,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Idempotency Key"
          },
          "moves": {
            "items": {
              "$ref": "#/components/schemas/OrganizeMoveIn"
            },
            "maxItems": 200,
            "minItems": 1,
            "title": "Moves",
            "type": "array"
          },
          "path": {
            "maxLength": 4096,
            "minLength": 1,
            "title": "Path",
            "type": "string"
          },
          "volume_id": {
            "minimum": 1.0,
            "title": "Volume Id",
            "type": "integer"
          }
        },
        "required": [
          "volume_id",
          "path",
          "moves"
        ],
        "title": "OrganizePlanRequest",
        "type": "object"
      },
      "OrganizeProposalOut": {
        "description": "A reviewable Organize suggestion for one folder (the model proposes; the server clamps).",
        "properties": {
          "considered": {
            "title": "Considered",
            "type": "integer"
          },
          "items": {
            "items": {
              "$ref": "#/components/schemas/OrganizeItemOut"
            },
            "title": "Items",
            "type": "array"
          },
          "model": {
            "title": "Model",
            "type": "string"
          },
          "rejected": {
            "title": "Rejected",
            "type": "integer"
          },
          "root": {
            "title": "Root",
            "type": "string"
          },
          "volume_id": {
            "title": "Volume Id",
            "type": "integer"
          }
        },
        "required": [
          "root",
          "volume_id",
          "model",
          "considered",
          "rejected",
          "items"
        ],
        "title": "OrganizeProposalOut",
        "type": "object"
      },
      "OrganizeSuggestRequest": {
        "description": "Ask for a proposed reorganisation of the files under one folder (read-only).",
        "properties": {
          "max_files": {
            "default": 60,
            "maximum": 200.0,
            "minimum": 1.0,
            "title": "Max Files",
            "type": "integer"
          },
          "path": {
            "maxLength": 4096,
            "minLength": 1,
            "title": "Path",
            "type": "string"
          },
          "volume_id": {
            "minimum": 1.0,
            "title": "Volume Id",
            "type": "integer"
          }
        },
        "required": [
          "volume_id",
          "path"
        ],
        "title": "OrganizeSuggestRequest",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    }
  },
  "info": {
    "description": "HTTP API for Fathomline (built on the Fathom engine). The read/query routes are RBAC- and scope-gated; the agent ingest routes authenticate by mTLS client-cert fingerprint behind the proxy; the remediation, preview and deployment surfaces are default-OFF and individually gated. See the project docs for auth and deployment.",
    "summary": "The storage-estate analyzer — read/query surface, agent ingest, and the default-OFF write/deploy surfaces.",
    "title": "Fathomline API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/organize/activity": {
      "get": {
        "description": "Summarise recent churn under ``path`` — a read-only \"re-organise?\" hint (ADR-021 Phase 3).\n\nReads the incremental change feed only; nothing is moved and nothing is auto-applied. Gated by\n``organize_enabled`` + ``VIEW_METADATA`` + the server-authoritative scope, like /suggest.",
        "operationId": "organize_activity_api_v1_organize_activity_get",
        "parameters": [
          {
            "in": "query",
            "name": "volume_id",
            "required": true,
            "schema": {
              "minimum": 1,
              "title": "Volume Id",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "path",
            "required": true,
            "schema": {
              "maxLength": 4096,
              "minLength": 1,
              "title": "Path",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "since_hours",
            "required": false,
            "schema": {
              "default": 24,
              "maximum": 720,
              "minimum": 1,
              "title": "Since Hours",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OrganizeActivityOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Organize Activity",
        "tags": [
          "organize"
        ]
      }
    },
    "/api/v1/organize/plan": {
      "post": {
        "description": "Build + persist a reversible MOVE plan from an approved subset (no FS touch; ADR-023).\n\nDefault-OFF behind BOTH gates; idempotent on ``idempotency_key``. The persisted ``plan_id`` is\nthen run through the remediation dry-run/execute spine (which carries the signed-job, drift, MFA\nand blast-cap guards). This route only *builds* — it dispatches nothing.",
        "operationId": "organize_plan_api_v1_organize_plan_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OrganizePlanRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OrganizePlanOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Organize Plan",
        "tags": [
          "organize"
        ]
      }
    },
    "/api/v1/organize/suggest": {
      "post": {
        "description": "Propose a reorganisation for the files under ``path`` (read-only; nothing is moved).",
        "operationId": "organize_suggest_api_v1_organize_suggest_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OrganizeSuggestRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OrganizeProposalOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Organize Suggest",
        "tags": [
          "organize"
        ]
      }
    },
...
  }
}